updated mistake upload file

+++ server/src/express/controllers/based-schemas/user.js
@@ -39,15 +39,18 @@ export const handleUpdateAvatar = catchAsyncErrors(async (req, res, next) => {
     try {
       // remove old avatar
       await deleteFile(ref);
-      user.avatar = avatar;
-      await user.save();
     } catch (error) {
       // remove new image if error
-      await deleteFile(avatar.ref);
+      // remove new image if error
+      if (avatar.ref) {
+        await deleteFile(avatar.ref);
+      }
       const msg = 'Lỗi cập nhật ảnh đại diện. Vui lòng thử lại';
       return next(new ErrorHandler(500, msg, 4036));
     }
   }
+  user.avatar = avatar;
+  await user.save();
   res.json({
     success: true,
     message: 'Cập nhật ảnh đại diện thành công',
